Move resolver tests into a separate module
